submitAI/requirements.txt
browser-use==0.1.40
langchain-openai==0.3.1
langchain-anthropic==0.3.3
fastapi==0.110.0
uvicorn==0.29.0
jinja2==3.1.3
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]>=2.10.4
pydantic-settings>=2.0.0
python-dotenv==1.0.1
aiohttp>=3.9.5
celery==5.3.6
redis==5.0.2
pyjwt==2.8.0
